I have done modifications to serializers.py and views.py of the farmers package.  I removed a class that is no longer used by Django and wasn't being used in the project.
I modified the serializer to allow for changes to the models used
by Django 1.6.10

The biggest change in this commit is the requirements.txt file.
In this file the number of requirements has increased, but most
of the requirements are just dependencies of the core actually
being used.  The requirements.txt file was automatically
generated using the pip-dump command from the pip-tools package.