I removed some depracated methods from the models file.  This still requires a bit of testing but from what I see everything still works.
Note that django-registration needs to be removed from your library
for this to work.  Try using pip uninstall django-registration.
django registration does not have support for 1.7 and now the
django-registration-redux is used.  It has the same functionality
as previous versions.  django-registration-redux is a fork of the
django-registration because the app is no longer maintained by the
owner.

Please do a syncdb or migrate before launching!

Matthew Budram