Ok please note that the changes in this commit will affect the loading of the application!  The application now attempts to read the values from specific environment variables and if it doesn't find them it will complain and throw a 'KeyError'.
In this commit you will notice that there is a template file
provided.  The values in the file must be changed to match that of
the environment that it is being run.  It also is important to note
that these are the exact environmental variables that must be exported
for the heroku instance.

Also to keep a nicely managed environment I have used virtualenv.
This allows me to keep my machine free from harvest specific environ-
mental variables and have it running it it's own sandbox.

Please remember this will break the application in it's current form!

Matthew Budram